+++ examples/platformer/js/entities/coin.js
@@ -17,8 +17,6 @@ game.CoinEntity = me.Sprite.extend({
         this.body.addShape(new me.Ellipse(this.width / 2, this.height / 2, this.width, this.height))
         // set the collision type
         this.body.collisionType = me.collision.types.COLLECTABLE_OBJECT;
-        // enable physic
-        this.isKinematic = false;
     },
 
     /**