+++ src/entity/entity.js
@@ -133,9 +133,6 @@
 
             // disable for entities
             this.autoTransform = false;
-
-            // enable collision detection
-            this.isKinematic = false;
         },
 
         /** @ignore */