+++ src/physics/body.js
@@ -277,6 +277,9 @@
                 }
                 this.updateBounds();
             }
+
+            // automatically enable physic when when a body is added to a renderable
+            this.ancestor.isKinematic = false;
         },
 
         /**