+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Physic : me.game.world is now a specific container (based on me.Container) that will manage all related physic updates
 - Physic : gravity for physic bodies is now defined globally under me.game.world.gravity, and can be scaled individually using me.Body.gravityScale
 - Physic : added a rotate method to me.Body to allow rotating all the body shapes properly (by default around the body bounding box center)
+- Physic : melonJS will now automatically enable body physic when when a me.Body object is added to any renderable
 - Pointer : fix an issue with uncorrect horizontal delta value for wheel event
 - Pointer : all registered events for a specific renderable are now automatically released upon the object destruction
 - Renderable : add support for free-texture-packer (http://free-tex-packer.com), the free Texture Packer alternative (thanks @PLAYERKILLERS)